perf(activity): use purecomponent for activity list item and extract it
